Reorganize code in almost one file per type
I know this might not be the Rust way of doing thing, but at least for
me it is much easy to maintain.
