+++ ghost_engine/src/resources/resource_creation_error.rs
@@ -1,0 +1,4 @@
+#[derive(Debug, PartialEq)]
+pub enum ResourceCreationError {
+    AlreadyRegistered,
+}
